Fix PHPStan issues
After a few updates in PHPStan, a few errors started to pop up. This
commit fixes those.

+++ library/Rules/Hetu.php
@@ -50,7 +50,7 @@ final class Hetu extends Simple
             return false;
         }
 
-        $id = $day . $month . $year . $individualNumber;
+        $id = (int) ($day . $month . $year . $individualNumber);
         $validationKeys = str_split('0123456789ABCDEFHJKLMNPRSTUVWXY');
 
         return $validationKeys[$id % 31] === $controlCharacter;
